refactor: fix some refactoring bugs
