+++ frontend/app/_apis/api-test.test.ts
@@ -1,6 +1,6 @@
 import { describe, afterEach, it, vi, expect } from "vitest";
 
-import apiTest from "@/apis/api-test"; // Adjust the path accordingly
+import apiTest from "@/_apis/api-test"; // Adjust the path accordingly
 
 describe("apiTest", () => {
     afterEach(() => {
